using size_t for count arguments

+++ includes/msh_structs.h
@@ -7,7 +7,7 @@
 
 struct command {
 	char	**argv;
-	int	argc;
+	size_t	argc;
 };
 
 #endif /* MSH_STRUCTS_H */
